+++ src/billboard/billboard_barrel.c
@@ -24,8 +24,8 @@ BillboardEnemyUpdate billboard_update_barrels(const PlayerState *player) {
     (void)player;
 
     // Deactivation mutates the target registry, so iterate a stable snapshot.
-    u8 snapshot[BILLBOARD_OBJECT_COUNT];
-    const u8 *target_indices = billboard_registry_target_indices();
+    u16 snapshot[BILLBOARD_OBJECT_COUNT];
+    const u16 *target_indices = billboard_registry_target_indices();
     const u16 target_count = billboard_registry_target_count();
     for (u16 i = 0; i < target_count; i++) snapshot[i] = target_indices[i];
